修正

+++ fm/sc/include/unit.hpp
@@ -325,7 +325,7 @@ public:
 
     //! @brief ベクトル量の大きさ
     Element magnitude() const
-        {return Element(std::sqrt(_data[0]*_data[0] + _data[1]*_data[1] + _data[2]*_data[2]));}
+        {return Element(std::sqrt(double(_data[0]*_data[0] + _data[1]*_data[1] + _data[2]*_data[2])));}
 };
 
 //! @brief ベクトルの単項+
